pyproject.toml
[tool.poetry]
name = "yamlconfigloader"
version = "0.1.0"
description = ""
authors = ["Denis Odinokov <denis.odinokov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
ruamel-yaml = "^0.18.6"
pydantic = "^2.8.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
